LPS-123794 | LPS-123795 Applies review suggestions
* Use parentField property on useDrop hook as a optional property properly
* Removes unused code and leftovers of DataConverter migration
* Removes unnecessary getDropHandler file. Since We're calling Form actions on the new useDrop hook, this file is unnecessary.
* It also fixes a overTarget error when dragging a field on Table View
* Instead of DIRECTIONS as a constant for use on resize, use RESIZE_DIRECTIONS
* Simplifies ResizableColumn responsibilities and props traversing. See: https://github.com/liferay-forms/liferay-portal/pull/132#discussion_r581354059
* Remove DraggableField component. This component is unnecessary since it just contain a simple markup, doesn't handle logic and can cause a bad idea of handling some Dragging logic.
* Move FieldDragPreview to a separated file
* Avoids losing the reference of mousemove and mouseup handlers when rendering using useCallback. See: https://github.com/liferay-forms/liferay-portal/pull/132#discussion_r581323376
* Move ResizableColumn component to a separated file
* Removes DRAG_DATA_DEFINITION_FIELD_MOVE const
* Removes DRAG_FIELDSET_MOVE dragType since We have the same implementation of DRAG_FIELD_TYPE_MOVE on useDrop
* Fixes drag/drop behaviours on fieldset. It fixes a weird behaviour of dragging a field inside a fieldset. Also, It adds some displayNames for forwardedRef components
* Avoid surpress eslint for fieldDragPreview effect. When using `ref.current` directly to mutate and remove on unmount there was an error but when storing this ref on a variable inside the effect it will not lose the reference :)
