+++ modules/apps/app-builder/app-builder-web/src/main/resources/META-INF/resources/js/pages/form-view/CustomObjectFieldsList.es.js
@@ -123,8 +123,8 @@ const getFieldTypes = ({
 			disabled: DataLayoutVisitor.containsField(dataLayoutPages, name),
 			dragAlignment: 'right',
 			dragType: isFieldGroup
-				? DragTypes.DRAG_FIELDSET_MOVE
-				: DragTypes.DRAG_DATA_DEFINITION_FIELD_MOVE,
+				? DragTypes.DRAG_FIELDSET_ADD
+				: DragTypes.DRAG_DATA_DEFINITION_FIELD_ADD,
 			icon: fieldTypeSettings.icon,
 			isCustomField: !customProperties.nativeField,
 			isFieldSet,
@@ -263,21 +263,6 @@ const CustomObjectFieldsList = ({keywords}) => {
 	const showCategories =
 		!!customFieldTypes.length && !!nativeFieldTypes.length;
 
-	const getDataDefinitionField = (fieldName) => {
-		const dataDefinitionField = dataDefinition.dataDefinitionFields.find(
-			(field) => field.name === fieldName
-		);
-
-		const settingsContext = dataLayoutBuilder.getDDMFormFieldSettingsContext(
-			dataDefinitionField
-		);
-
-		return {
-			...dataDefinitionField,
-			settingsContext,
-		};
-	};
-
 	const fieldTypeListProps = {
 		deleteLabel: Liferay.Language.get('delete-from-object'),
 		keywords,
@@ -299,7 +284,6 @@ const CustomObjectFieldsList = ({keywords}) => {
 				{...fieldTypeListProps}
 				dataDefinition={dataDefinition}
 				fieldTypes={customFieldTypes}
-				getDataDefinitionField={getDataDefinitionField}
 				showEmptyState={false}
 			/>
 
@@ -313,7 +297,6 @@ const CustomObjectFieldsList = ({keywords}) => {
 				{...fieldTypeListProps}
 				dataDefinition={dataDefinition}
 				fieldTypes={nativeFieldTypes}
-				getDataDefinitionField={getDataDefinitionField}
 				showEmptyState={false}
 			/>
 		</>
