+++ modules/apps/app-builder/app-builder-web/src/main/resources/META-INF/resources/js/pages/table-view/DropZonePlaceholder.es.js
@@ -65,7 +65,7 @@ const getStyle = (container, index, total) => {
 
 const Placeholder = ({container, index, onAddFieldName, total}) => {
 	const [{canDrop, overTarget}, drop] = useDrop({
-		accept: DragTypes.DRAG_FIELD_TYPE_MOVE,
+		accept: DragTypes.DRAG_FIELD_TYPE_ADD,
 		collect: (monitor) => ({
 			canDrop: monitor.canDrop(),
 			overTarget: monitor.isOver(),
