+++ modules/apps/data-engine/data-engine-taglib/src/main/resources/META-INF/resources/data_layout_builder/js/drag-and-drop/dragTypes.es.js
@@ -12,9 +12,7 @@
  * details.
  */
 
-export const DRAG_DATA_DEFINITION_FIELD_MOVE = 'dataDefinitionField:move';
 export const DRAG_FIELD_TYPE_MOVE = 'fieldType:move';
-export const DRAG_FIELDSET_MOVE = 'fieldset:move';
 
 export const DRAG_DATA_DEFINITION_FIELD_ADD = 'dataDefinitionField:add';
 export const DRAG_FIELD_TYPE_ADD = 'fieldType:add';
